Move code from worker to state.
